% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{compute_age}
\alias{compute_age}
\title{derive age from dates}
\usage{
compute_age(start_date, end_date = NA, placeholder_date = "2020-12-31")
}
\arguments{
\item{start_date}{Date object for starting date of interval (typically birthdate)}

\item{end_date}{Optional Date object for ending date of interval
such as death date or simulation end date. If NA, will use current
date of the session.}
}
\value{
age in years
}
\description{
Calculate age using starting date and optional end date
}
