% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{compute_default_coordinates}
\alias{compute_default_coordinates}
\title{obtain min and max mapping coordinates}
\usage{
compute_default_coordinates(con)
}
\description{
obtain min and max mapping coordinates
}
