% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simclindata.shiny-package.R
\docType{package}
\name{simclindata.shiny}
\alias{simclindata.shiny}
\alias{simclindata.shiny-package}
\title{A Shiny App package}
\description{
Read more about the workshop at [https://rpharma.shinyprod.com](https://rpharma.shinyprod.com/).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rpodcast/simclindata.shiny}
  \item Report bugs at \url{https://github.com/rpodcast/simclindata.shiny/issues}
}

}
\author{
\strong{Maintainer}: Eric Nantz \email{theRcast@gmail.com}

}
